# [platform] improve perf of PropertyChangedEventArgsExtensions