--- a/Xamarin.Forms.Platform.iOS/CollectionView/PropertyChangedEventArgsExtensions.cs
+++ b/Xamarin.Forms.Platform.iOS/CollectionView/PropertyChangedEventArgsExtensions.cs
@@ -1,25 +1,48 @@
 ﻿using System.ComponentModel;
+using System.Runtime.CompilerServices;
 
 namespace Xamarin.Forms.Platform.iOS
 {
 	internal static class PropertyChangedEventArgsExtensions
 	{
+		[MethodImpl(MethodImplOptions.AggressiveInlining)]
 		public static bool Is(this PropertyChangedEventArgs args, BindableProperty property)
 		{
 			return args.PropertyName == property.PropertyName;
 		}
 
-		public static bool IsOneOf(this PropertyChangedEventArgs args, params BindableProperty[] properties)
+		[MethodImpl(MethodImplOptions.AggressiveInlining)]
+		public static bool IsOneOf(this PropertyChangedEventArgs args, BindableProperty p0, BindableProperty p1)
 		{
-			foreach (BindableProperty property in properties)
-			{
-				if (args.PropertyName == property.PropertyName)
-				{
-					return true;
-				}
-			}
+			return args.PropertyName == p0.PropertyName ||
+				args.PropertyName == p1.PropertyName;
+		}
 
-			return false;
+		[MethodImpl(MethodImplOptions.AggressiveInlining)]
+		public static bool IsOneOf(this PropertyChangedEventArgs args, BindableProperty p0, BindableProperty p1, BindableProperty p2)
+		{
+			return args.PropertyName == p0.PropertyName ||
+				args.PropertyName == p1.PropertyName ||
+				args.PropertyName == p2.PropertyName;
+		}
+
+		[MethodImpl(MethodImplOptions.AggressiveInlining)]
+		public static bool IsOneOf(this PropertyChangedEventArgs args, BindableProperty p0, BindableProperty p1, BindableProperty p2, BindableProperty p3)
+		{
+			return args.PropertyName == p0.PropertyName ||
+				args.PropertyName == p1.PropertyName ||
+				args.PropertyName == p2.PropertyName ||
+				args.PropertyName == p3.PropertyName;
+		}
+
+		[MethodImpl(MethodImplOptions.AggressiveInlining)]
+		public static bool IsOneOf(this PropertyChangedEventArgs args, BindableProperty p0, BindableProperty p1, BindableProperty p2, BindableProperty p3, BindableProperty p4)
+		{
+			return args.PropertyName == p0.PropertyName ||
+				args.PropertyName == p1.PropertyName ||
+				args.PropertyName == p2.PropertyName ||
+				args.PropertyName == p3.PropertyName ||
+				args.PropertyName == p4.PropertyName;
 		}
 	}
 }