@ahmedalejo 2020-01-04
This makes lots of sense 

I always concluded that property handling has always being Achilles hill of MVVM. some resolve this by holding a lazily initialized HashTable for the properties to compare with and having the property name comparison to be of _O_(1) complexity. On the down side it **could** use more memory.

In your case, using overloads reduces memory allocations which is a great fit for mobile.