@jonathanpeppers 2020-01-06
Hi @ahmedalejo,

The problem with this, is it's going to do this work for every `Label` instance:

```csharp
public class Label
{
    private HashSet<string> fontRelatedProperties;

    public HashSet<string> FontRelatedProperties => 
        fontRelatedProperties ??= new[] { "Font", "FontSize", "FontFamily", "TextColor", "FontAttribute"}.ToHashSet();
```

I suspect the act of creating the `HashSet` would be slower than the time saved on property changes, but I would have to benchmark it to be sure.

It is probably also not worth making the property lazy, because every `Label` will have at least one property change. Unless you have an app with a bunch of blank invisible `Label`'s? 😄 

Side note, you should just do this so it don't allocate an intermediate `string[]`:

```csharp
new HashSet<string>
{
	"Font", "FontSize", "FontFamily", "TextColor", "FontAttribute"
};
```
I would not use this `ToHashSet` method at all.

I don't think Xamarin.Forms can take a dependency on Fody, but we are looking into what we can do at build time for improving runtime performance.