@ahmedalejo 2020-01-06
**Hi** @jonathanpeppers 
- The idea behind lazy was for controls that may not be used in an app(or at least not at startup) e.g CarouselPage, TableView.
- Making the `HashSet` static could do the trick for allocation per control per instance.