@hartez 2020-01-07
> I also found a couple places where IsOneOf was called with one
parameter. I think calling Is instead makes sense?

Yes. I think that's just folks copy/pasting and not thinking it through. Another nice thing about dropping `params` is that nobody can make that mistake again.