@hartez 2020-01-07
>  define up to 5 parameters. If you need more, please reexamine your life choices.

Oh, man, don't look at VisualElementTracker on iOS.