@hartez 2020-01-07
DROID-API27-FR-ISSUES failure is unrelated to these changes.